DR2-1583 Add file logging.
I can't convince podman to not write the timestamp all over my nice json
and my original plan of having a systemd service which writes a cleaned
log isn't working because systemd decides to spawn multiple processes so
we end up with multiple logs lines.

Rather than trying to work out what's up with it, I've changed the
logging to log to a file. I'll then mount a volume from the host to the
/poduser/logs directory and we should get logs written directly from the
code.