Fix vulnerabilities in image
The openjdk image is behind with its version of alpine and there were a
good few fixes in the newer versions so I've replaced it with the base
alpine image and installed the jre manually. Running apk upgrade gets
the latest versions of packages which usually clears a few CVEs as well.