DR2-1605 Purge corrupted objects.
It is possible to end up with a corrupted object if SIGKILL is sent
while the OCFL library is writing.

If this happens, we now check for the corrupted object exception, purge
the corrupted version from the repository and rethrow the exception.
This will cause the message to be picked up again and this time, after
the purge, the write should complete successfully.

I've added in a test for it.