Increase max heap
There is an IO in preservica with 1200 assets which has 1800ish versions
and a 45Mb inventory file.

The OCFL library reads the whole thing into memory and it's running out
of heap space.

I have no idea what to set it to but this should cover us for a while.