DR2-1843 Stop using Option for SO
The CC process is only interested in IO and CO messages as there’s nothing we want to store about an SO.

The incoming message is returned as `Option[ReceivedSnsMessage]` where None indicates an SO.

The problem with this is that we then do `distinctBy` on this value which flattens all the SO values so that we only get the first one.

So when multiple SO messages come in, all but one are not being deleted.

The problem is mostly caused by using an `Option` to represent an SO message. I’ve realised now that this is silly and I should have an SoReceivedMessage case class to represent this.

We also weren't using the messageText parameter other than for deduping and we can use the ref for that as the odds for having the same uuid for a CO and an IO are small enough to ignore.