DR2-1890 Use mutable head and other fixes.
There are a few changes here

The first is the change to use the FIFO queue and the mutable head.
There is a description in the README as to how this works.

The second is to filter out any event actions from Preservica for
command_download. When this process downloads the file for the first
time, it will create another two versions because the download is appended to
the event actions and this changes the checksum of both the IO and CO
metadata. This is a bit silly.

The third is that there's another bug which I thought I'd fixed.
We're deduplicating messages based on the id which then selects a random
message where there are duplicates. We then delete that message but the
message/receipt handle that was removed by `distinctBy` stays in the
queue until it is processed without any duplicates.