DR2-1924 Custodial copy errors
We were getting this exception

`io.ocfl.api.exception.ObjectOutOfSyncException: Failed to update mutable HEAD of object 12f27ed2-920f-4cb0-b938-ee0a8c8cb6a1. Changes are out of sync with the current object state.\n\tat io.ocfl.core.storage.DefaultOcflStorage.createRevisionMarker(DefaultOcflStorage.java:979)\n\tat`

I'm not 100% sure of the cause.
I originally thought it was that the commit call wasn't acquiring the semaphore before committing which was causing a
race condition somewhere but adding the semaphore didn't get rid of the
error. That being said, it's probably a nice thing to have.

The other thing that it could be is that we were creating a new
VersionInfo object and passing that in. That does have an effect on the
mutable head version I think which is probably what was causing us
problems. I've passed in null instead and the error seems to have gone.

The most serious problem was that if there was an error inside the
processor as there was here, the semaphore wasn't being released so the
process would just hang.

I've added a condition that releases the semaphore if there are any
errors in any of the fibers. I can't really add a test for it without
some hefty refactoring but I've tried it on uat and it works.