Release on error
As part of my ongoing effort to find out why the cc process keeps
freezing up I'm trying this.

I suspect there's an error somewhere in the OCFL processes that isn't
being caught or reported or something so I've added this in to release
the semaphore on error.

It might help, it might not, it's not made it any worse anyway.