DR2-2103 Scan docker images with Wiz
I've added this as an sbt build task to avoid having to list each docker
image we build.

I've also added this into the test workflow so that it will scan a local
copy of the built image so we catch antything before its merged.

Once it has been merged to main, we scan it again and then use with Wiz
CLI tag once it has been pushed to ECR.