DR2-2070 Exclude embedded files
Preservica will now not return a download url in the metadata if there
is no bitstream. So we check if there is a url. If there isn't, we don't
download the file. If there's no file, we don't try to stage it.

I've updated the tests and added a couple of new ones.