DR2-2127 Logging for CC
This logs the source ID in the json object for each message so we can
search for it.

It also logs the receipt handle when we delete the messages.

I've changed the docker image so that the log4j2.properties is now on
the docker container instead of inside the jar. This means that we can
mount a directory on the host to the container directory and change the
logging level if we need to, for example, setting the preservica client
to debug level.