DR2-2430 Make deletions work in CC.
Because we don't know the entity type of a deleted entity, the original
way of doing this does't work. So all the checks about which entity
types can be deleted have gone.
We now check to see if the id is in the OCFL repository, if it's not, we
ignore it and move on as this will be a CO or SO.
If it's a deleted entity, we don't send a message to the SNS topic for a
deleted message as this only goes to the db builder anyway.