Ignore empty checksums
There is one file (an potentially more in the future) which is an
extracted attachment from an email which doesn't have a checksum in
Preservica.

There's not much we can do about this so we're ignoring things that have
no checksum.

The risk that something has no checksum but isn't one of these extracted
files is acceptable. Every file in the OCFL repo has a checksum as well.