DR2-2580 Conditional check on update.
We are getting instances, we’re not sure how where the CC confirmer
 is processing a message after the row has been deleted from Dynamo.

This means that the row is inserted without the input field which
the state change handler is expecting. The state change handler then
fails along with anything else that’s in that current batch.

Adding a conditional check should fix this. We don't care if the check
fails, that just means that the asset has already been checked so we
ignore that specific error.